REFACTOR(Contracts.Validators): changed namespace
